+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - OpenGLEngine times startup and shutdown steps and reports time spent
 - sphere, cylinder, cone, disk, torus supports normal mapping
 - OBJ models can auto-generate tangents for normal mapping
+- Fixed bug for shader introspection displaying uniform array names
 
 ## v 5.25.0.0 - 09 Dec 2025
 - MD5Models support normal mapping